Windowクラスのコンストラクタに親Windowを指定できるように対応。Window生成時初期ボーダースタイルを明示的に設定するように対応。ワーニング修正。

+++ visual/win32/BasicDrawDevice.cpp
@@ -715,8 +715,8 @@ void TJS_INTF_METHOD tTVPBasicDrawDevice::EndBitmapCompletion(iTVPLayerManager *
 	float dt = (float)( DestRect.top < ClipRect.top ? ClipRect.top : DestRect.top );
 	float dr = (float)( DestRect.right > ClipRect.right ? ClipRect.right : DestRect.right );
 	float db = (float)( DestRect.bottom > ClipRect.bottom ? ClipRect.bottom : DestRect.bottom );
-	float dw = DestRect.get_width();
-	float dh = DestRect.get_height();
+	float dw = (float)DestRect.get_width();
+	float dh = (float)DestRect.get_height();
 
 	// ‚Í‚İo‚µ‚Ä‚¢‚é•‚ğ‹‚ß‚é
 	float cl = dl - (float)DestRect.left;
